requirements.txt
Flask>=3.0.0
Waitress>=3.0.0
bcrypt>=4.1.0
Pillow>=10.0.0
markdown>=3.5.0
nanoid>=2.0.0
